[iOS] FEAT: Session을 프로토콜화 하여 DIP #208
